wheel exception - czy w klasie abstract wheel mogę zostawić abstract hub? Zamiast określać już na tym etapie czy ma to być front czy rear?

+++ src/main/java/p76/bicycles/service/compatibility/CompatibilityWheelService.java
@@ -102,9 +102,9 @@ public class CompatibilityWheelService {
 
     public List<Integer> tyreRimRange(Wheel wheel) {
         try {
-            int rim = wheel.getRim().getInnerWidth();
-            int temp = 0;
-            for (int key : dataService.diameterMap().keySet()) {
+            double rim = wheel.getRim().getInnerWidth();
+            double temp = 0;
+            for (double key : dataService.diameterMap().keySet()) {
                 if (rim == key) {
                     temp = key;
                 }
